Fixed conf_dir and pkg_dev on debian
There should be a better way of doing this, as I'm hardcoding the version name into the config dir.

On Debian based systems, postgresql is configured in "clusters" using some Debian specific tools. The main cluster is called ... main, which is why the `conf_dir` has a suffix of `main`.