fix(postgres.manage): only reload modules if needed (fix: #214)
We don't want the `postgres-reload-modules` state reporting any changes
to the user unless there is a need to reload the modules. We should only
need to reload the modules if the state that manages installing the
client libraries makes any changes (ie. new packages installed).

In turn, the states created by the `format_state` macro only need to
ensure that any module reload occurs before those states are applied. In
which case using a `require` makes more sense that using `onchanges`
(which may not be guaranteed to trigger).